requirements.txt
streamlit
openai
python-dotenv
selenium 
webdriver-manager